Fix: Incorrect input to file-grep-regex-action
path -> filename

This bug has been present since the initial code drop, but only
applied to setup.py projects, which haven't received extensive
testing to date.

This update also resequences the extraction code in action.yaml.
It's possible for Python projects to have BOTH a pyproject.toml
file and a setup.py, for example, where a Python project needs to
use setuptools in conjunction with C code that must be compiled.

I've added a second project to the action's tests in testing.yaml.
That repository was how this bug was detected.

Signed-off-by: Matthew Watkins <mwatkins@linuxfoundation.org>